// types/location/place/place-type.type.ts
import type { Place } from "./place";

export interface PlaceType {
  id: string;

  // ======================================================
  // BASIC
  // ======================================================

  name: string;

  nativeName?: string | null;

  code: string;

  slug: string;

  description?: string | null;

  // ======================================================
  // MEDIA
  // ======================================================

  icon?: string | null;

  thumbnail?: string | null;

  // ======================================================
  // DISPLAY
  // ======================================================

  sortOrder: number;

  // ======================================================
  // STATUS
  // ======================================================

  active: boolean;

  // ======================================================
  // RELATIONS
  // ======================================================

  places: Place[];

  // ======================================================
  // TIMESTAMP
  // ======================================================

  createdAt: Date;

  updatedAt: Date;
}